CC-CVRP/requirements.txt
vrplib==1.3.0
taichi==1.6.0
numpy==1.26.1
dimod==0.12.13
dwave.system==1.21.0
networkx==2.8.8
matplotlib==3.8.0


